No trial field run seemed to go okay. A few minor global variables like seeds, ranks and RNG streams might still require tweaking though

+++ production_c/force.c
@@ -108,7 +108,7 @@ int Gauge_force(double *dSdpi,Complex *u11t, Complex *u12t, int *iu, int *id, fl
 int Force(double *dSdpi, int iflag, double res1, Complex *X0, Complex *X1, Complex *Phi,Complex *u11t, Complex *u12t,\
 		Complex_f *u11t_f,Complex_f *u12t_f,int *iu,int *id,Complex gamval[5][4],Complex_f gamval_f[5][4],\
 		int gamin[4][4],double *dk4m, double *dk4p, float *dk4m_f,float *dk4p_f,Complex_f jqq,\
-		float akappa,float beta,int *ancg){
+		float akappa,float beta,double *ancg){
 	/*
 	 *	Calculates dSds at each intermediate time
 	 *
